code/requirements.txt
faiss-cpu>=1.10.0
jupyter>=1.1.1
langchain-anthropic>=0.3.10
langchain-community>=0.3.20
langchain-mcp-adapters>=0.0.7
langchain-openai>=0.3.11
langgraph>=0.3.21
mcp>=1.6.0
notebook>=7.3.3
pymupdf>=1.25.4
python-dotenv>=1.1.0
streamlit>=1.44.1 
nest-asyncio>=1.6.0
qdrant-client
PyPDF2
python-docx
pandas
openpyxl
torch
transformers
tiktoken
python-multipart
aiofiles
pdfplumber
python-pptx
xlrd
fastapi
litellm>=1.67.1
lingua-language-detector>=2.1.0
semchunk>=3.2.1
regex>=2024.11.6
openai>=1.75.0
pymongo
pymupdf4llm